fix: add prop on ExperienceRoot to use lazy loading or script injecting the visual editor
