+++ packages/experience-builder-sdk/src/blocks/VisualEditor.tsx
@@ -1,35 +1,34 @@
-import React, { useEffect } from 'react';
-import { InternalSDKMode } from '../types';
+import React, { useEffect, useState } from 'react';
+import { InternalSDKMode, VisualEditorMode } from '../types';
 import { VisualEditorContextProvider } from './editor/VisualEditorContext';
-import { VISUAL_EDITOR_CONTAINER_ID } from '@contentful/experience-builder-types';
 
-const version = '0.0.1-pre-20231213T103332.0';
-const scriptUrl = `https://unpkg.com/@contentful/experience-builder-visual-editor@${version}/dist/renderApp.js`;
-
-const VisualEditor: React.FC = () => {
-  useEffect(() => {
-    const script = document.createElement('script');
-    script.type = 'module';
-    script.src = scriptUrl;
-    script.async = true;
-
-    document.body.appendChild(script);
-
-    return () => {
-      document.body.removeChild(script);
-    };
-  }, []);
-
-  return <div id={VISUAL_EDITOR_CONTAINER_ID} />;
-};
-
-type VisualEditorRootProps = {
+export type VisualEditorRootProps = {
   initialLocale: string;
   mode: InternalSDKMode;
+  visualEditorMode: VisualEditorMode;
 };
 
-export const VisualEditorRoot: React.FC<VisualEditorRootProps> = ({ initialLocale, mode }) => {
-  // in editor mode locale can change via sendMessage from web app, hence we use the locale from props only as initial locale
+export const VisualEditorRoot: React.FC<VisualEditorRootProps> = ({
+  initialLocale,
+  mode,
+  visualEditorMode,
+}) => {
+  const [VisualEditor, setVisualEditor] = useState<React.ComponentType | null>(null);
+
+  useEffect(() => {
+    if (visualEditorMode === 'lazyLoad') {
+      import('@contentful/experience-builder-visual-editor').then((module) => {
+        setVisualEditor(() => module.default);
+      });
+    } else {
+      import('./InjectVisualEditor').then((module) => {
+        setVisualEditor(() => module.default);
+      });
+    }
+  }, [visualEditorMode]);
+
+  if (!VisualEditor) return null;
+
   return (
     <VisualEditorContextProvider mode={mode} initialLocale={initialLocale}>
       <VisualEditor />
