+++ packages/experience-builder-types/src/types.ts
@@ -176,6 +176,8 @@ export type CompositionTree = {
 export type ExternalSDKMode = 'preview' | 'delivery';
 export type InternalSDKMode = ExternalSDKMode | 'editor';
 
+export type VisualEditorMode = 'lazyLoad' | 'injectScript';
+
 /**
  * Internally defined style variables are prefix with `cf` to avoid
  * collisions with user defined variables.
